requirement.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
idna==2.10
oauthlib==3.1.1
python-telegram-bot==13.7
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
TeslaPy==1.3.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
